feat: добавляет webpack сборку, разделяет функции на компоненты
