+++ public/index.html
@@ -41,6 +41,6 @@
           </section>
       </div>
     </main>
-    <script src="../src/main.js"></script>
+    <script src="/js/bundle.js"></script>
   </body>
 </html>
